LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigation-back-same-document-preventDefault-expected.txt


FAIL navigation.back() same-document preventDefault assert_equals: expected "http://localhost:8800/navigation-api/navigate-event/navigation-back-same-document-preventDefault.html#" but got ""
